Update charge information dynamically with Vue.js. (#6)
We use Vue.js to bind the number of tickets and use that information
when Stripe's checkout form is opened by Vue's callback. The
chargable amount is updated in the buy button as well.

This change also introduces an app state for vue, sets the public Stripe key
and loads the event price from the database.